requirements.txt
streamlit==1.40.1
tensorflow==2.12.0
pillow==10.4.0
streamlit-option-menu